New R Script Using TDA to Analyze DJIA tickers
WIP R script to replicate the results in "Topology Data Analysis of Critical Transitions in Financial Networks" by Marian Gidea.

Gidea's paper says that we can use this method to detect early signs of critical transitions in the market. He applied this to the 2008 financial crisis. We can apply this to the COVID crash or cryptocurrencies.

To-do: Documentation on result interpretation and some further work on the code. This is just preliminary work. Not all tickers can be generated, which makes results differ slightly than those in the paper.